+++ autogpt_platform/frontend/src/app/(platform)/admin/spending/page.tsx
